Fix "assertion !intr_context() failed" error. All passed.
In the external interrupt context, do not use thread_yield, but
intr_yield_on_return, because thread_yield should be invoked only
when the running thread is not in the external interrupt context.

1. Unnecessary:

Although the document requires that the priority argument to
thread_create should be ignored, it is not necessary for mlfqs to
explicitly disable priority initialization when a thread is created,
because later priority update dose not depend on the initial
priority value. It depends only on recent_cpu and nice value.

2. Required, but no effect to tests:

This current implementation sets nice and recent_cpu value
to zero in init_thread. However, the initial thread starts with
a nice value of zero, and other threads start with a nice
value inherited from their parent thread. Likewise, the initial
value of recent_cpu is zero in the first thread created, or
the parent's value in other new threads.

TODO:
  - Fix 1 and 2.