Modify `kill' to exit with -1 status.
Changed `thread_exit' into `exit(-1)'.

When a user process attempts to read or write a memory at kernel address
or an address that is not mapped, the process should be terminated
with a -1 exit code.
These invalid pointer accesses cause a page fault and the handler
finally invokes a `kill' function, which terminates the current thread.

Passed:
  - make tests/userprog/bad-read.result
  - make tests/userprog/bad-write.result
  - make tests/userprog/bad-read2.result
  - make tests/userprog/bad-write2.result
  - make tests/userprog/bad-jump.result
  - make tests/userprog/bad-jump2.result