Add the process hierarchy and exec/wait system.
Passed:
  - make tests/userprog/exec-once.result
  - make tests/userprog/exec-arg.result
  - make tests/userprog/exec-bound.result
  - make tests/userprog/exec-bound-2.result
  - make tests/userprog/exec-bound-3.result
  - make tests/userprog/exec-multiple.result
  - make tests/userprog/exec-missing.result
  - make tests/userprog/exec-bad-ptr.result
  - make tests/userprog/wait-simple.result
  - make tests/userprog/wait-twice.result
  - make tests/userprog/wait-killed.result
  - make tests/userprog/wait-bad-pid.result
  - make tests/userprog/multi-recurse.result

Added `process' structure which is malloc'ed in `start_process'
and managed in the similar way of "shared_ptr" in C++11: If
a process block is neither owned by a thread nor referenced by
parent, it is safely `free'ed.

Every information related to process hierarchy and exec/wait
status is maintained in the `struct process', not in thread'.

Therefore, a "thread" can freely exits without considering
"process" deallocation: "process" can be alive without owner
thread or without parent thread. But without both the owner and
parent, the process will be deallocated.