Implement lazy loading.
Previously, executable code and data segments are immediately loaded
in memory during process setup.
We modified the implementation so that only a stack is immediately
loaded.

Because executable code and data segments are initially not present
pages, a not-present page fault should occur when a process accesses
these unloaded segments. The kernel dose not load these segments until
the fault occurs, that is, the page fault handler loads the code/data
segments if demanded, and resumes the process's execution.

In order for the page fault handler to know how to load the fault page,
the process must create SPTEs and set the necessary information.
Thus, we modified load_segment() defined in userprog/process.c so that
it creates some SPTEs which contain supplemental information, without
allocating any physical frame.

The supplemental page table (SPT) and its entries (SPTEs) are managed
by vm/page.c.

Previously, allocating a physical frame is done by palloc_get_page()
defined in threads/palloc.c. Instead, we modified the implementation so
that frame_alloc() defined in vm/frame.c performes the frame allocation
in order to easily support frame eviction.

FRAME EVICTION IS NOT IMPLEMENTED YET.