Add frame eviction.
Passed:
  - make tests/vm/page-linear.result
  - make tests/vm/page-parallel.result
  - make tests/vm/page-merge-seq.result
  - make tests/vm/page-merge-par.result
  - make tests/vm/page-merge-stk.result
  - make tests/vm/page-shuffle.result

For the sake of convenience and efficiency, FTE and SPTE are doubly
linked. That is, if a physical frame is allocated and mapped to a
user virtual page, a FTE corresponding to the frame and a SPTE
corresponding to the user page point to each other.

Overall, FTEs are all connected to some set of SPTEs, but there exists
some SPTEs to which FTE is not connected. It is because there is
a limitation on the number of physical frames that can be allocated.

IMPORTANT:
The frame eviction algorithm was implemented in FIFO way. We have to
reimplement the algorithm so that it approximates LRU; the algorithm
should perform at least as well as the simple variant of the "second
chance" or "clock" algorithm.