Add mmap() and munmap() system calls. One error left.
Passed:
  - make tests/vm/mmap-read.result
  - make tests/vm/mmap-close.result
  - make tests/vm/mmap-unmap.result
  - make tests/vm/mmap-overlap.result
  - make tests/vm/mmap-twice.result
  - make tests/vm/mmap-write.result
  - make tests/vm/mmap-exit.result
  - make tests/vm/mmap-shuffle.result
  - make tests/vm/mmap-bad-fd.result
  - make tests/vm/mmap-clean.result
  - make tests/vm/mmap-inherit.result
  - make tests/vm/mmap-misalign.result
  - make tests/vm/mmap-null.result
  - make tests/vm/mmap-over-code.result
  - make tests/vm/mmap-over-data.result
  - make tests/vm/mmap-over-stk.result
  - make tests/vm/mmap-remove.result
  - make tests/vm/mmap-zero.result

Fail:
  - make tests/vm/page-merge-mm.result

We just added two functions that add or remove some set of SPTEs
which are used to load mmap'ed user virtual pages; sys_mmap() and
sys_unmap(). That is, if SPTEs are well allocated to the mmap region,
page loading is handled by the already implemented lazy loading
procedure.

A strange error remains. Obviously, we called the sys_mmap_exit()
function so that all mmap mappings are munmap'ed at the end of the
process (see process_exit() defined in userprog/process.c), but we
failed the test "page-merge-mm.c" which exits without explicit
munmap() calls.