Update frame eviction algorithm.
We modified frame_get_victim() so that it performs "clock" algorithm.

However, some test failed:
  - make tests/vm/page-parallel.result

This test sometimes passed, but sometimes failed.
We guessed it's probably because of the synchronization problem. In
addition, this test is related to the frame eviction itsef, not the
algorithm; testing of previous FIFO implementation showed similar
behaviors. So, we will focus on the critical section that might be
generated during frame eviction.