Add synchronization during frame eviction.
Passed:
  - make tests/vm/page-parallel.result

When removing a SPTE, process must wait until the SPTE's frame
eviction completes, if being processed, and then remove it
if possible.

During the eviction of physical frame, the corresponding FTE is
pinned, that is, between the time the FTE was pinned inside the
function frame_get_victim() and the last time it was unpinned
inside the function page_load().

It is unsafe to free or remove SPTE's allocated frame, or even
itself, without waiting:

Suppose that a frame is evicted from SPTE SRC of process P1 to
DST of P2. See frame_do_eviction(). To deprive SRC of its FTE and
physical frame, P2 should reference SRC and SRC's FRAME member.
However, after context switch, if P2 removes and frees SRC and
allocated frame which is being evicted, whether by exit or explicit
call to page_remove_entry(), P2's referencing is invalid.

The solution is described in page_destruct_frame() and frame_free().