Change spin lock to Pintos's default lock.
Previously, we used spin lock with a boolean variable PINNED to
protect a critical section of frame eviction.

Our spin lock implementation consumed unnecessary thread ticks, and
thus we changed it to the Pintos's default lock.

In addition, for the sake of clarity, we changed the function name
of safe destructor from page_destruct_frame()
to wait_and_destruct_frame().