Add swap_free().
If some set of SPTEs of the exiting process indicates that their
contents reside in swap slots, then they must be freed; otherwise,
some swap slots are in "used" state and can never be freed.

For this purpose, we added a function swap_free() which just frees
the swap slot indicated by the given slot index.